Removed println. Separated files. Renamed SkinnyMessageHelper -> SkinnyMessageImplicits, and put it in skinny.mailer.implicits
