+++ mailer/src/main/scala/skinny/mailer/implicits/SkinnyMessageImplicits.scala
@@ -1,12 +1,13 @@
-package skinny.mailer
+package skinny.mailer.implicits
 
 import javax.mail.{ BodyPart, Message }
 import javax.mail.internet.{ MimeBodyPart, MimeMessage }
+import skinny.mailer.{ MimeBodyPartOps, MimeMessageOps, SkinnyMessage }
 
 /**
  * SkinnyMessageHelper provide implicit conversions
  */
-trait SkinnyMessageHelper {
+trait SkinnyMessageImplicits {
   import scala.language.implicitConversions
   implicit def mimeMessageOps[T >: SkinnyMessage <: Message](m: T) = new MimeMessageOps {
     override def instance = m match { case a: MimeMessage => a }
